requirements.txt
flask==2.3.3
streamlit==1.28.1
openai
langchain==0.1.20
langchain-openai==0.1.7
langchain-core==0.1.52
langchain-community==0.0.38
PyPDF2==3.0.1
requests==2.31.0
python-dotenv==1.0.0
faiss-cpu==1.7.4
tiktoken
pydantic==2.5.3